Add portfolio page and tailwindcss
Why:
I would like to have a personal website to display some of the things
i've done on the internet.

this commit:
adds a portfolio page as the root page and moves the current root page
to a new Doodles page. Tailwindcss was added to implement the styles as
it works well for non-designers and static sites.