Make doodles responsive and use tailwind
Why:
Currently all of the doodles are not responsive to screen sizes and look
bad on smaller screen sizes. We'd like for the doodles to look okay on
all screen sizes.

This commit updates all fo the doodles to be responsive and replaces
most of the non-tailwind css with tailwind classes