Introduce elm
Why:
I'd like to play around with elm

This commit:
Introduces elm to the project and places a Browser.sandbox in a new html
page: squares.html with a corresponding Main.elm. I've added a build
script to build the elm app to dist/squares_main.js. It would be nice in
a future commit to consolidate the build targets so there is just the
one main.js again, but this is a quick way to add elm without needing to
change anything else in the code base.